src/backend/requirements.txt
# 기본 의존성 (OpenAI 기반)
openai>=1.0.0
python-dotenv>=1.0.0

# Vector DB 관련
qdrant-client>=1.7.0
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
numpy>=1.24.0

# 한국어 처리
kss>=4.5.0

# 기타 유틸리티
requests>=2.31.0
typing-extensions>=4.8.0